feat: reuse PortifolioTable between Trade and Portifolio pages + props to filter content
