+++ apps/dex/src/compounds/Margin/Portfolio.tsx
@@ -1,70 +1,11 @@
 import type { NextPage } from "next";
 
-import { Suspense } from "react";
-import { useRouter } from "next/router";
-import clsx from "clsx";
-import dynamic from "next/dynamic";
-import Link from "next/link";
-
-const DEFAULT_OPTION_ITEM = "open-positions";
-const OPTIONS_ITEMS = [
-  {
-    title: "Open Positions",
-    slug: "open-positions",
-    content: dynamic(() => import("~/compounds/Margin/OpenPositions"), {
-      loading: () => (
-        <div className="bg-gray-850 p-10 text-center">Loading...</div>
-      ),
-    }),
-  },
-  {
-    title: "History",
-    slug: "history",
-    content: dynamic(() => import("~/compounds/Margin/History"), {
-      loading: () => (
-        <div className="bg-gray-850 p-10 text-center">Loading...</div>
-      ),
-    }),
-  },
-];
+import { PortfolioTable } from "~/compounds/Margin/PortfolioTable";
 
 const Portfolio: NextPage = () => {
-  const router = useRouter();
-  const activeOption =
-    (router.query["option"] as string) || DEFAULT_OPTION_ITEM;
-  const currentTab = OPTIONS_ITEMS.find((item) => item.slug === activeOption);
-  const TabContent = currentTab?.content || null;
-
   return (
     <section className="mt-4 border border-gold-800 rounded overflow-hidden">
-      <ul className="flex flex-row text-sm bg-gray-800 rounded-tl rounded-tr">
-        {OPTIONS_ITEMS.map(({ title, slug }) => {
-          const isTabActive = currentTab?.slug === slug;
-          return (
-            <li key={slug}>
-              <Link href={{ query: { ...router.query, option: slug } }}>
-                <a
-                  className={clsx(
-                    "flex mx-4 py-3",
-                    isTabActive ? "text-white border-current" : "text-gray-400",
-                  )}
-                >
-                  {title}
-                </a>
-              </Link>
-            </li>
-          );
-        })}
-      </ul>
-      {TabContent && (
-        <Suspense
-          fallback={
-            <div className="bg-gray-850 p-10 text-center">Loading...</div>
-          }
-        >
-          <TabContent />
-        </Suspense>
-      )}
+      <PortfolioTable />
     </section>
   );
 };
